requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
pikepdf>=8.0.0
pandas>=2.0.0
pdfplumber>=0.11.0
